BACKEND

+++ service/api/utils/getPostPhotoPath.go
@@ -4,6 +4,6 @@ import (
 	"fmt"
 )
 
-func GetPostPhotoPath(user_id int, post_id int) string {
+func GetPostPhotoPath(user_id int, post_id int64) string {
 	return fmt.Sprintf("./storage/%d/posts/%d.jpeg", user_id, post_id)
 }
